Refactor copy to recruitment cycle to Service Object
